attempt to redirect to heroku

+++ app/controllers/queue_albums_controller.rb
@@ -14,8 +14,7 @@ class QueueAlbumsController < ApplicationController
   def destroy
     queue_album = QueueAlbum.find_by(id: params[:id])
     user_id = queue_album[:user_id]
-    # list_id_for_update = list_album[:list_id]
-    if queue_album
+    if queue_album && session[:user_id] == user_id
       queue_album.destroy
       updated_user = User.find_by(id: user_id)
       updated_queue = updated_user.queue_albums
